app/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-mysql-rds==0.4.0
djangorestframework==3.11.2
pytz==2021.1
sqlparse==0.4.1
